breakup hyprpaper file into modules
